Refactor: 셀렉션 생성 관련 리포지토리 / 서비스 리펙토링
- DB 접근 관련 함수는 리포지토리에 보관
